Reorganize the forms module
Put all foo.{html,css,spec.ts,ts} files into the 'foo' directory
