+++ client/app/forms/component-mapper/component-mapper.service.spec.ts
@@ -1,11 +1,11 @@
 
 import { ComponentMapperService } from './component-mapper.service';
-import { AutocompleteControlComponent } from './controls/autocomplete-control.component';
-import { CheckboxControlComponent } from './controls/checkbox-control.component';
-import { DateTimeControlComponent } from './controls/date-time-control.component';
-import { EnumeratedControlComponent } from './controls/enumerated-control.component';
-import { InputControlComponent } from './controls/input-control.component';
-import { FormControlType } from './form-control-spec.interface';
+import { AutocompleteControlComponent } from '../dynamic-controls/autocomplete-control.component';
+import { CheckboxControlComponent } from '../dynamic-controls/checkbox-control.component';
+import { DateTimeControlComponent } from '../dynamic-controls/date-time-control.component';
+import { EnumeratedControlComponent } from '../dynamic-controls/enumerated-control.component';
+import { InputControlComponent } from '../dynamic-controls/input-control.component';
+import { FormControlType } from '../form-control-spec.interface';
 
 describe('ComponentMapperService', () => {
     let mapper: ComponentMapperService;
