+++ client/app/forms/component-mapper/component-mapper.service.ts
@@ -1,10 +1,10 @@
 import { Injectable, Type } from '@angular/core';
-import { AutocompleteControlComponent } from './controls/autocomplete-control.component';
-import { CheckboxControlComponent } from './controls/checkbox-control.component';
-import { DateTimeControlComponent } from './controls/date-time-control.component';
-import { EnumeratedControlComponent } from './controls/enumerated-control.component';
-import { InputControlComponent } from './controls/input-control.component';
-import { FormControlType } from './form-control-spec.interface';
+import { AutocompleteControlComponent } from '../dynamic-controls/autocomplete-control.component';
+import { CheckboxControlComponent } from '../dynamic-controls/checkbox-control.component';
+import { DateTimeControlComponent } from '../dynamic-controls/date-time-control.component';
+import { EnumeratedControlComponent } from '../dynamic-controls/enumerated-control.component';
+import { InputControlComponent } from '../dynamic-controls/input-control.component';
+import { FormControlType } from '../form-control-spec.interface';
 
 /**
  * Maps FormControlTypes to component types. Exists as a service instead of a
